move _colormap and _dtype
